Switch Cues to use float for timecodes

+++ pisync/lib/api/create_cue_request.py
@@ -5,5 +5,5 @@ class CreateCueRequest(BaseModel):
 
     name: str
     sourceMediaId: int
-    sourceMediaTimecode: int
+    sourceMediaTimecode: float
     targetMediaId: int
